+++ src/OpenTelemetry.Resources.ProcessRuntime/ProcessRuntimeDetector.cs
@@ -20,38 +20,33 @@ internal sealed class ProcessRuntimeDetector : IResourceDetector
     public Resource Detect()
     {
         var frameworkDescription = RuntimeInformation.FrameworkDescription;
-        string netRuntimeName;
+        string netRuntimeVersion = string.Empty;
 #if NETFRAMEWORK
         var netFrameworkVersion = GetNetFrameworkVersionFromRegistry();
-#endif
-
-        var lastSpace = frameworkDescription.LastIndexOf(' ');
-        if (lastSpace != -1)
+#pragma warning disable CS8602 // Dereference of a possibly null reference.
+        string netRuntimeName = ".NET Framework";
+#pragma warning restore CS8602 // Dereference of a possibly null reference.
+        if (netFrameworkVersion != null && !string.IsNullOrEmpty(netFrameworkVersion))
         {
-            // sample result '.NET Framework 4.8.9195.0'
-            netRuntimeName = frameworkDescription.Substring(0, lastSpace);
-#if NETFRAMEWORK
-            netFrameworkVersion ??= frameworkDescription.Substring(lastSpace + 1);
-#endif
+            var lastSpace = netFrameworkVersion.LastIndexOf(' ');
+            netRuntimeVersion = netFrameworkVersion.Substring(lastSpace + 1);
         }
         else
         {
-            // do not expect to be here, all checked implementation has common FrameworkDescription format - '{Name With Optional Spaces} {Version}'
-            netRuntimeName = "unknown";
-#if NETFRAMEWORK
-            netFrameworkVersion ??= "unknown";
-#endif
+            netRuntimeVersion = Environment.Version.ToString();
         }
+#else
+        netRuntimeVersion = Environment.Version.ToString();
+#pragma warning disable CA1307 // Specify StringComparison for clarity
+        string netRuntimeName = frameworkDescription.Replace(netRuntimeVersion, string.Empty).Trim();
+#pragma warning restore CA1307 // Specify StringComparison for clarity
+#endif
 
         return new Resource(
         [
             new(ProcessRuntimeSemanticConventions.AttributeProcessRuntimeDescription, frameworkDescription),
             new(ProcessRuntimeSemanticConventions.AttributeProcessRuntimeName, netRuntimeName),
-#if NETFRAMEWORK
-            new(ProcessRuntimeSemanticConventions.AttributeProcessRuntimeVersion, netFrameworkVersion),
-#else
-            new(ProcessRuntimeSemanticConventions.AttributeProcessRuntimeVersion, Environment.Version.ToString()),
-#endif
+            new(ProcessRuntimeSemanticConventions.AttributeProcessRuntimeVersion, netRuntimeVersion),
         ]);
     }