+++ test/OpenTelemetry.Resources.Process.Tests/ProcessDetectorTests.cs
@@ -13,10 +13,25 @@ public class ProcessDetectorTests
         var resource = ResourceBuilder.CreateEmpty().AddProcessDetector().Build();
 
         var resourceAttributes = resource.Attributes.ToDictionary(x => x.Key, x => x.Value);
+#if NET
+        Assert.Equal(11, resourceAttributes.Count);
+#else
+        Assert.Equal(10, resourceAttributes.Count);
+#endif
 
-        Assert.Equal(2, resourceAttributes.Count);
-
+        Assert.IsType<long>(resourceAttributes[ProcessSemanticConventions.AttributeProcessArgsCount]);
+        Assert.IsType<string[]>(resourceAttributes[ProcessSemanticConventions.AttributeProcessCommandArgs]);
+        Assert.IsType<string>(resourceAttributes[ProcessSemanticConventions.AttributeProcessCommandLine]);
+        Assert.IsType<string>(resourceAttributes[ProcessSemanticConventions.AttributeProcessExecName]);
+#if NET
+        Assert.IsType<string>(resourceAttributes[ProcessSemanticConventions.AttributeProcessExecPath]);
+#endif
+        Assert.IsType<bool>(resourceAttributes[ProcessSemanticConventions.AttributeProcessInteractive]);
         Assert.IsType<string>(resourceAttributes[ProcessSemanticConventions.AttributeProcessOwner]);
         Assert.IsType<long>(resourceAttributes[ProcessSemanticConventions.AttributeProcessPid]);
+
+        Assert.IsType<string>(resourceAttributes[ProcessSemanticConventions.AttributeProcessStartTime]);
+        Assert.IsType<string>(resourceAttributes[ProcessSemanticConventions.AttributeProcessTitle]);
+        Assert.IsType<string>(resourceAttributes[ProcessSemanticConventions.AttributeProcessWorkingDir]);
     }
 }